Update BaseProtocol to return int in LAYOUT_SIZE

+++ engine/src/main/java/br/angarion/dev/engine/communication/BaseProtocol.java
@@ -17,7 +17,7 @@ public final class BaseProtocol {
         )
     );
 
-    public static final long LAYOUT_SIZE = LAYOUT.byteSize();
+    public static final int LAYOUT_SIZE = (int) LAYOUT.byteSize();
 
     public static final VarHandle TOTAL_SIZE = LAYOUT.varHandle(
         PathElement.groupElement("totalSize")
